requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
flake8==7.1.1
fonttools==4.54.1
fpdf2==2.8.1
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
soupsieve==2.6
types-fpdf2==2.8.1.20241011
typing_extensions==4.12.2
urllib3==2.2.3
